+++ packages/manager/src/components/ColorPalette/ColorPalette.stories.mdx
@@ -7,8 +7,8 @@ import { ColorPalette } from './ColorPalette';
 
 Add a new color to the palette, especially another tint of gray or blue, only after exhausting the option of using an existing color.
 
-- Colors used in light mode are located in `themeFactory.ts`
-- Colors used in dark mode are located in `themes.ts`
+- Colors used in light mode are located in `foundations/light.ts`
+- Colors used in dark mode are located in `foundations/dark.ts`
 
 If a color does not exist in the current palette and is only used once, consider applying the color conditionally:
