+++ packages/manager/src/components/core/Grid.tsx
@@ -15,7 +15,7 @@ import { withStyles } from '@mui/styles';
 import { Theme } from '@mui/material/styles';
 import { Breakpoint } from '@mui/material';
 import type { GridProps } from '@mui/material/Grid';
-import { breakpoints } from 'src/themeFactory';
+import { breakpoints } from 'src/foundations/breakpoints';
 
 const SPACINGS = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10];
 const GRID_SIZES = ['auto', true, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12];
