+++ packages/manager/src/foundations/fonts.ts
@@ -1,0 +1,5 @@
+export const fonts = {
+  normal: '"LatoWeb", sans-serif',
+  semiBold: '"LatoWebSemibold", sans-serif',
+  bold: '"LatoWebBold", sans-serif',
+} as const;
